cleanup jaccard generation
